Remove OagRepository abstraction
